pyproject.toml
[tool.poetry]
name = "async_fastapi_jwt_auth"
version = "0.6.4"
description = "Async fork of FastAPI extension that provides JWT Auth support (secure, easy to use and lightweight)"
authors = ["Yan Khachko <a@slnk.icu>"]
readme = "README.md"
license = "MIT"
classifiers = [
  "Environment :: Web Environment",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
  "Topic :: Software Development :: Libraries :: Python Modules"
]

[tool.poetry.dependencies]
python = ">=3.8"
fastapi = ">=0.110.0"
PyJWT = ">=2.8.0"
httpx = ">=0.23.3"
pydantic-settings = ">=2.2.1"
cryptography = ">=42.0.5"

[tool.poetry.extras]
asymmetric = ["cryptography"]

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
poethepoet = "^0.21.1"
pytest = "7.4.0"
uvicorn = "^0.23.2"
pytest-asyncio = "^0.21.1"
coveralls = "==3.3.1"
mkdocs = ">=1.1.2,<2.0.0"
mkdocs-material = ">=5.5.0,<6.0.0"
markdown-include = ">=0.5.1,<0.6.0"
isort = "^5.12.0"

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.poe.tasks]
autoformat.sequence = [
    {cmd = "black async_fastapi_jwt_auth tests examples"},
    {cmd = "isort async_fastapi_jwt_auth tests examples"},
]

test.sequence = [
  {cmd = "pytest"}
]

[tool.poetry.urls]
Documentation = "https://sijokun.github.io/async-fastapi-jwt-auth"
Source = "https://github.com/sijokun/async-fastapi-jwt-auth"
